project1/requirements.txt
streamlit 
fastapi
fastapi[standard] 
uvicorn[standard]
dvc 
dvc[azure]
azure-storage-blob
pyjwt
passlib

pillow==10.4.0
transformers==4.45.1
python-dotenv==1.0.1
MarkupSafe==2.1.5 
jinja2==3.1.4 
mpmath==1.3.0 
networkx==3.3 
sympy==1.13.3 
torch==2.4.1 
torchvision==0.19.1
python-multipart==0.0.10
bcrypt==3.2.2
streamlit==1.38.0
requests==2.32.3

